requirements.txt
opencv-python
matplotlib
numpy
odc-stac
pandas
rioxarray
Pillow
planetary-computer
pystac-client
geopy